requirements.txt
geopy
requests
pyserial
folium
pywebview